git-svn-id: https://vaio/svn/vssdb@324 bf672a44-3a04-1d4a-850b-c2a239875c8c

+++ Comic/ComicTest/ComicTestDlg.cpp
@@ -24,7 +24,7 @@ CComicTestDlg::CComicTestDlg(CWnd* pParent /*=NULL*/)
 	, m_y(0)
 	, m_x1(0)
 	, m_x2(1)
-	, m_step(0.1)
+	, m_step((float)0.1)
 {
 	m_hIcon = AfxGetApp()->LoadIcon(IDR_MAINFRAME);
 	pMath = NULL;
